+++ index.html
@@ -5,7 +5,6 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <script src="/js/main.js"></script>
   <title>Кексобукинг</title>
 </head>
 <body>
@@ -229,6 +228,8 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
+
+  <script src="/js/main.js" type="module"></script>
 
 </body>
 </html>
